Rethinking Recursive Structure With PlayNodes
Thinking about using a 'PlayNode' which is essentially a single decision point from the perspective of any given player. That player has options of cards to play and figures out what is optimal by recursively creating new 'PlayNodes' from the perspective of the next player after they play a card.