Improved genOpponentHands() logic
Added check that a player cannot have a card of a suit that they have already not followed. For example if Hearts are led in round #1, then the player must follow hearts if he has any. If instead the player played Spades, then any future random hand generation cannot be hearts, since we know the player must not have hearts.