New ideas on hand generation
Thinking about efficient ways to generate hands for opponents based on what they bid. Obviously once a player bids we know more about their hand than it just being any random cards, but its difficult to implement that in code. The genHands() function in DecisionPoint is a work in progress to figure out how to do that logic.