Implemented smart genOpponentHands
genOpponentHands now checks whether an opponent has already bid. If so, it guarantees that the hand generated will be one where the optimal bid is equal to the bid the opponent previously made. If the opponent has not bid, it just generates a random hand as normal. This works as long as no cards have been played in a game (hence the commented out test that includes one card played prior to genOpponentHands being called)