Modify makePlay to cut out upon card play that gets bonus
If a card played will get you the bonus, you'll obviously want to play that card. There is an issue where two cards could get you the bonus, so there is code to randomize the order in which the cards are tested so that each will be chosen over a series of runs.