Update makeBid to run until it is confident with an answer
makeBid ran with one set of randomly generated opponent hands and returned the optimal bid for just that scenario. It now loops for x number of randomly generated opponent hands until the optimal bids returned from each run are summed up and clearly point to one bid being optimal. For example, if you run makeBid 10 times and it says to bid 2 every time, you can confidently bid 2. If instead is says to bid 2 6 times and to bid 1 4 times, you probably want to run more simulations to find which is optimal. makeBid now runs until it is happy that it has found the best bid.